backend/requirements.txt
asgiref==3.2.10
atomicwrites==1.4.0
attrs==21.4.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.10
colorama==0.4.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==36.0.1
defusedxml==0.7.1
Django==2.2.19
django-extensions==3.1.5
django-extra-fields==3.0.2
django-filter==2.4.0
django-templated-mail==1.1.1
djangorestframework==3.11.2
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
flake8==4.0.1
flake8-broken-line==0.4.0
flake8-isort==4.1.1
flake8-plugin-utils==1.3.2
flake8-polyfill==1.0.2
flake8-return==1.1.3
gunicorn==20.0.4
idna==3.3
importlib-metadata==1.7.0
iniconfig==1.1.1
isort==5.10.1
itypes==1.2.0
Jinja2==3.0.3
MarkupSafe==2.0.1
mccabe==0.6.1
oauthlib==3.1.1
packaging==21.3
pep8-naming==0.12.1
Pillow==9.0.1
pluggy==0.13.1
psycopg2-binary==2.8.6
py==1.11.0
pycodestyle==2.8.0
pycparser==2.21
pyflakes==2.4.0
PyJWT==2.1.0
pyparsing==3.0.7
pytest==6.2.4
pytest-django==4.4.0
pytest-pythonpath==0.7.3
python-dotenv==0.19.1
python3-openid==3.2.0
pytz==2021.3
reportlab==3.6.6
requests==2.27.1
requests-oauthlib==1.3.0
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.2.0
sqlparse==0.4.2
testfixtures==6.18.3
toml==0.10.2
uritemplate==4.1.1
urllib3==1.26.8
zipp==3.7.0
